refactor: add createVerifyProofResultSuccess, createVerifyProofResultFailure to helpers and move getPublicDataFromProofs and createTrustedDataFromProofData to helpers
